requirements.txt
flask==3.0.0
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
python-dotenv==1.0.0
joblib==1.3.2
gunicorn==21.2.0
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
